refactor(telegram-menu-structured): Refactor MenuItem class to handle command checking and value setting
This commit refactors refresh  and  adds a condition to refresh the MenuItem and its nested items when the command to check starts with the MenuItem's command or when the MenuItem is the root, to avoir refresh MenuITem's on other branches of menu tree.
Additionally it updates the logic for removing the 'lastCommand' value when the command to check is 'cmdExit' and the MenuItem is the root.